codegen: zero-pad under-sized fixed-length array initializers
C's "the rest default to zero" rule (a braced init with fewer elements
than the array's declared length) was only handled for the
char-array-from-string idiom. render_scalar_or_nested_array now pads
any literal-dim array the same way via std::mem::zeroed() - real
corpus cases: g_game.c's pars[4][10] ({0} rows), wi_stuff.c's
lnodes[NUMEPISODES][NUMMAPS]. Found via the real --emit-rust run
(cargo build -p doom_rs: 783 -> 780 errors).

+++ doom_rs/src/hu_stuff.rs
@@ -532,7 +532,7 @@ pub static mut english_shiftxform: [std::ffi::c_char; 128] = unsafe {
     ]
 };
 
-pub static mut frenchKeyMap: [std::ffi::c_char; (128) as usize] = unsafe {
+pub static mut frenchKeyMap: [std::ffi::c_char; 128] = unsafe {
     [
         0,
         1,
