+++ doom_rs/src/tables.rs
@@ -84,7 +84,7 @@ pub unsafe extern "C" fn SlopeDiv(num: std::ffi::c_uint, den: std::ffi::c_uint) 
     todo!("body not yet translated")
 }
 
-pub static mut finetangent: [std::ffi::c_int; (4096) as usize] = unsafe {
+pub static mut finetangent: [std::ffi::c_int; 4096] = unsafe {
     [
         (-(170910304)),
         (-(56965752)),
@@ -4185,7 +4185,7 @@ pub static mut finetangent: [std::ffi::c_int; (4096) as usize] = unsafe {
     ]
 };
 
-pub static mut finesine: [std::ffi::c_int; (10240) as usize] = unsafe {
+pub static mut finesine: [std::ffi::c_int; 10240] = unsafe {
     [
         25,
         75,
@@ -14430,7 +14430,7 @@ pub static mut finesine: [std::ffi::c_int; (10240) as usize] = unsafe {
     ]
 };
 
-pub static mut tantoangle: [angle_t; (2049) as usize] = unsafe {
+pub static mut tantoangle: [angle_t; 2049] = unsafe {
     [
         0, 333772, 667544, 1001315, 1335086, 1668857, 2002626, 2336395, 2670163, 3003929, 3337694,
         3671457, 4005219, 4338979, 4672736, 5006492, 5340245, 5673995, 6007743, 6341488, 6675230,
